refactor: remover configuração desnecessária de header
O header 'content-type' é implicitamente setado pelo express, sem
necessidade de configuração manual.
